Fixes an FTBFS when a mix of PIC and PIE compiler flags are used
On some platforms the linker fails when the PIC and PIE are both used.It
was discovered on an 64 bit environment although it worked on an 32 bit
environment.
The Qt5Widgets_EXECUTABLE_COMPILE_FLAGS flags should be used when building
an executable.
The library is built with the PIC flag although the PIC flag
is not explicitly set. The add_library() command, used with the SHARED
option, adds CMAKE_SHARED_LIBRARY_CXX_FLAGS variable to the build rules. The
CMAKE_SHARED_LIBRARY_CXX_FLAGS variable contains the appropriate Position
Independent Code (PIC) flags to the platform.

Reported-by: Alexis López Zubieta <azubieta@estudiantes.uci.cu>
Tested-by: Alexis López Zubieta <azubieta@estudiantes.uci.cu>
Signed-off-by: Luís Pereira <luis.artur.pereira@gmail.com>