Fixes an FTBFS on the test target
It failed to build the test target when the library wasn't already
installed. It failed when the xdgdesktopfile and xdgdirs headers weren't
system wide accessible.

The library headers included should be from the library that is being
built. Using:
    #include "qtxdg/xdgdesktopfile.h"
lead to the already installed headers were being used.
Adding ${CMAKE_SOURCE_DIR} to the include list and removing the qtxdg/
prefix from the #include makes sure that we are using our local header.
Tested with in source build and out of source build.

Reported-by: Alexis López Zubieta <azubieta@estudiantes.uci.cu>
Tested-by: Alexis López Zubieta <azubieta@estudiantes.uci.cu>
Signed-off-by: Luís Pereira <luis.artur.pereira@gmail.com>