Put iconloaderfix into own namespace
In Qt5.2 it looks like a private property of QIconTheme went missing so the classes no longer line up.  This causes a VERY odd crash when you load an icon via the XdgIcon::fromTheme(...) since it tries to copy the QIconTheme and m_contentDirs is missing.  The order changed so memory is 7 colors of screwed up.

This patch puts the fixed loader into its own QtXdg namespace to keep the classes isolated and intended operation correct.