Allows QIconLoader to search in multiple theme directories.
Even when QIcon::themeSearchPaths() returns a list QIconLoader doesn't
honor it. It will search for an icon in the first directory it finds (that
has the current theme).
QIcon still doesn't support a theme that is spread across several
directories. And it's a fundamental feature needed to support user
installed icons, e.g., the hicolor theme.