Adds a list of special non detachable execs
Some executables (pkexec) don't work with QProcess::startDetached().
Instead of calling QProcess::start(), which in some situations can lead
to the executable being killed as an result of his parent kill, we maintain
a list of executables that don't cope with QProcess::startDetached() and
call selectivily QProcess::start() / QProcess::startDetached().

By using this ugly solution we have the best of the two worlds.