Adds platform-data handling
From the standard: "All methods take a platform-data argument that is used
in a similar way to how environment variables might be used. Currently,
only one field is defined by the specification: desktop-startup-id. This
should be a string of the same value as would be stored in the
DESKTOP_STARTUP_ID environment variable, as specified by theStartup
Notification Protocol Specification."

http://standards.freedesktop.org/desktop-entry-spec/desktop-entry-spec-latest.html#dbus