build: Create separate Qt5XdgIconLoader target
After separating the iconloader engine there were no mean how to
use the qt5xdgiconloader separately in cmake projects (if we exclude
the use of generated pgkconfig file).
This commit adds generation of Qt5XdgIconLoader target and it's *.cmake
files. Also gives the Qt5XdgIconLoader as a dependency to Qt5Xdg.