qiconloader: Reuse Qt implementation
1. Use as much of the Qt QIconLoader implementation as possible (do not
just copy-paste all of the logic from Qt sources).
2. Do not store the currently set theme name in XdgIconLoader instance,
but reuse the underlying QIconLoader for it.

With this change there is only one place where the current theme (name,
paths, etc.) is stored => the calls XdgIcon::*themeName and
QIcon::*themeName are interchangable.

TODO: deprecate and/or remove the XdgIcon & XdgIconLoader wrapper
functions (which only call QIcon & QIconLoader functions).