Optimize QIconLoader::findIconHelper()
Profiling QIconLoader::findIconHelper() shows that a significant portion of CPU time is being spent in
QDir::exists(), which creates a new QFileInfo object for the sole purpose of determining whether the passed-in
file path is relative or absolute, and then calls QFile::exists(). In this context, we can just as easily
generate the absolute path and call QFile::exists() directly, avoiding the creation of extra QDir and QFileInfo
objects.

Change-Id: Ib0b4568b6c16d423eb6c1b15158e44ff141e6175
Task-number: QTBUG-46767
Reviewed-by: Gunnar Sletta <gunnar@sletta.org>

Port of:
https://code.qt.io/cgit/qt/qtbase.git/commit/src/gui/image/qiconloader.cpp?id=36aaf851ff2814e9e5c024e21b866c403137ff26

Note: Didn't port to our enhancements to the icon loader. They are not in
a loop, so the performance improvements are insignificant.