xdgiconloader: Correct hierarchy of fallbacks
This is my idea of a *practically* correct hierarchy of fallbacks, in
contrast to what XDG specification tells us.

The icon is searched in the theme *and all of its parents*. Then, if it
is not found, the search is done for the first "dash fallback" in the
theme and all of its parents. And so on for the second, third,... "dash
fallbacks".