xdgiconloader: Add GTK cache revalidation
In case an icon is added and the GTK cache is refreshed the already
running Qt application does not get the information from GTK cache.
That's because the cache file is changed (modified or deleted+newly
created), but the mapped memory used to read the file is not affected.

We will re-map the cache file for such situations to properly use the
cache also after the content change.