Cover all KDE colors schemes for symbolic icons
In our support for KDE symbolic icons, we only considered `.ColorScheme-Text`. But some Breeze symbolic icons use KDE-specific schemes like `.ColorScheme-NegativeText`. Of course, we don't use such colors but even `.ColorScheme-Text` is something we've taken from KDE (it could be any string otherwise). So, it's better to cover all KDE text colors but set all of them to the same color.

Without this patch, some Breeze icons with KDE-specific schemes (like the 16-px `window-close.svg` or `application-exit.svg`) would always be pitch black -- even with dark Qt styles -- because their schemes are erased. But with this patch, they are like other icons: light on dark backgrounds and dark on light ones.

In short, this is needed if we want KDE icon sets to be really usable under LXQt too.