Sharp SVG icons with scale factors > 1
Fixes https://github.com/lxqt/libqtxdg/issues/178

The SVG icon engine gives a too big pixmap with scale factors > 1, so that the final drawing will not be sharp if the real size is used. This patch scales down the pixmap to the actual size in order to get a sharp SVG icon, especially with non-digit scale factors.