XdgDefaultApps: Adds methods to get the installed (special) applications
Special application are: web browser, email client and file manager.

Not every single application that supports a mimetype/protocol can be
considered to be a "special" application. As an example Qmmp (a media
player) supports inode/directory (for the playlist functionality) but it's
not a file manager. A solution is to use the Categories entries from the
Desktop Entry Specification in conjunction with the supported mimetypes.

Please note that we are not restricting the user ability to set a app as
the default, XdgDefault::set.....() methods aren't changed.
Special applications should receive special treatment. It will enhance the
user UI experience.