Consider XDG_DATA_HOME when computing desktop entry id
The standard doesn't refer to XDG_DATA_HOME. It should be considered, allowing
the id of an desktop entry located in XDG_DATA_HOME to be properly computed.
XDG_DATA_HOME is taken into consideration when searching for a desktop entry,
it make no sense not to consider it for computing id.