Add default terminal (#265)
* XdgDesktopFile: Add a public tryExec method

It should be public.
The functionality already existed in the private checkTryExec() function.
Code updated to use QStandardPaths::findExecutable(). It does a better job.

* Adds a default terminal implementation

xdg-utils doesn't provide a utility script to open the registered terminal
emulator. xdg-terminal exists since 2006 but it didn't make to the first
league. A Desktop environment should give it's user the power of choice. So
we implemented it.

The default terminal emulator is stored in qtxdg.conf files, which can be
stored in several locations. QSettings own search order is used.
Additionally, it is possible to define desktop environment-specific default
terminal in a file named desktop-qtxdg.conf where desktop is the name of
the desktop environment (taken from XDG_CURRENT_DESKTOP environment
variable). For example, ~/.config/lxqt-qtxdg.conf defines user specific
default terminal override for LXQt.

These desktop-specific overrides take precedence over the corresponding
non-desktop-specific file. If the desktop-specific file isn't found the
non-specific will *not* be used.