Workaround for random lack of icons in newly installed apps
This patch is a workaround for a random problem whose cause should be outside `xdgiconloader/xdgiconloader.cpp` — `xdgiconloader.cpp` seems correct to me in its original form.

The source of the problem is unknown to me, but it makes newly installed apps have `application-x-executable` as their icons in Main Menu.

The patch is a compromise: it consults GTK icon cache if valid; otherwise, it'll revalidate it and search in the directory hierarchy if the icon isn't found in it — which is the whole point of the patch.

IMHO, a DE that can't show the icons of newly installed apps has a serious UX issue.

Closes https://github.com/lxqt/libqtxdg/issues/226