Invalidate GTK icon cache as soon as icons are installed
Monitoring the GTK cache itself wasn't enough for showing real icons of newly installed apps (in Main Menu). Contrary to my expectations, the reason was that the cache was invalidated too late.

Here, the folder is monitored instead, and the cache is invalidated as soon as something happens inside the folder. It will be validated again when it's updated.

Closes https://github.com/lxqt/libqtxdg/issues/226