Fixed file association with symlink desktop entries
Previously, the (final) targets of symlink desktop entries were loaded and failed to meet the requirements of mime-type association. The patch solves the problem by loading the symlinks themselves.

I didn't find any side effect in the code or in practice. The only difference I've found is that, if a menu item corresponding to a symlink desktop entry is dragged and dropped, it'll be created as a symlink, not as its target — which isn't a problem and seems more natural to me.

Fixes https://github.com/lxqt/libqtxdg/issues/287 and closes https://github.com/lxqt/lxqt-config/issues/897